fix: create deal form component and fix bugs
