+++ src/schemas/dealSchema.js
@@ -1,12 +1,13 @@
-import { date, object, string } from "yup";
+import { date, number, object, string } from "yup";
 
 export const dealSchema = object({
   title: string().min(5,"Title need at least 5").required("Title is required"),
   category: string().required("Please select category"),
-  status:string().required("Please select status"),
+  deal_status:string().required("Please select status"),
   seller:string().required("Please select seller"),
+  max_participants: number().required("Please set the maximum participants"),
   description:string().optional(),
   start_at: date().required("Start Date required"),
   deadline: date().required("Deadline Date required"),
-  images: string().optional()
+  // images: string().optional()
 })
